Include quota helper in application controller
The CloudResourceQuotaHelper was originally only included in the
MiqRequestController for handling provisioning requests.  However, it appears
that there are other controllers that can access the provisioning workflow;
namely the VmCloudController.

When a controller other than MiqRequestController attempts to render
provisioning workflow pages, the user sees the error described in the linked bug
below.

Instead of trying to track down all of the controllers that somehow play a part
in provisioning requests, the CloudResourceQuotaHelper is now part of the base
application controller and available to all controllers.

https://bugzilla.redhat.com/show_bug.cgi?id=1190852