Create less vms and ems_events and where they're used.
Altered some of the setup code to ensure most/all of the assertions matched at least 1 record.

There's probably some duplicate tests in here and more moving of Factories to where used but this is a good start.

Before:
Finished in 16.56 seconds
Finished in 16.18 seconds

After:
Finished in 5.21 seconds
Finished in 5.28 seconds