Move Foreman managers into ExtManagementSystems.
This is needed because so many other things rely on EMSes that having
them separate makes the calling so much more complex.  This includes
things like workers, reporting, REST API, validations, scheduling, etc.

This commit may break reporting and/or introduce inconsistencies in
reporting, which will be fixed in follow up work.