UI changes to scale OpenStack Infrastructure Provider
This change presents a simple form for changing the number of nodes
assigned to each role. A call to update the stack is submitted
when the form is submitted.

This form could be improved, because having to calculate whether you
are exceeding the maximum number of hosts in your head is somewhat
cumbersome, but it is good enough as a first version.