Function called incorrectly with switched parameters
The function get_field takes the field name as the first parameter
and dialog_name as the second parameter. It was being incorrectly
called with dialog_name as the first parameter and field_name as
the second parameter.
This used to cause the valid values for the field to be an empty hash