by default fix_auth should upgrade v2 keys.
There is an optimization to only upgrade records
encrypted with a v1_key (not those already upgraded).

That optimization will no longer be enabled by default.
If a customer wants to skip v2 keys, then --no-v2 is provided.

Currently, when there are records encrypted with a bad v2 key, they are
not fixed and customers are confused.