Destroying an OpenStack VM should not set state attr
The state attribute on VMs is read only and used purely as a normalization layer
in order to display a common state for all different types of VMs.

When VMs are destroyed (or, otherwise acted upon), the appliance will set a
temporary power state in order to reflect the impending change that's been
submitted to the provider.  The change submitted to the provider is handled
largely asynchronously with any updates from the change coming in the form of
events received after the fact.

The temporary power state change when destroying OpenStack VMs was trying to set
the state attribute.  By changing that to the raw_power_state attribute, the
power state is updated and reflected in the read only state attribute.

https://bugzilla.redhat.com/show_bug.cgi?id=1212052