Do not derive "available" values if we don't have any usage values.
The lack of cpu or mem usage values implies that the target being
collected is either off, or not configured for collection.  In both
cases, collecting "allocated" values does not make sense.  If off, the
target will not be given those resources, so they are not really
available.  If not configured for collection, then we should not be
doing the derivation at all.

The circumstance for this situation occurs when normal C&U for a target
is not enabled, but storage C&U still occurs.  When storage C&U comes
along it calls process_derived_columns, but some of those derived columns
should not be calculated in that state.  If the normal C&U were to come
along later, then it would fill in the missing details.

https://bugzilla.redhat.com/show_bug.cgi?id=1038869
https://bugzilla.redhat.com/show_bug.cgi?id=1212164