Adding OpenstackInfra parsing of configurations files
Adding OpenstackInfra parsing of configurations files. OpenStack
configparser does not define exactly the format of real conf
files. The main exceptions are not alowing : delimiter, limited
naming of variables to \w and that we are parsing commented
lines that are ignored by python parser. These commented lines
contains descrtiption of each attribute and default value in
case the attribute is not defined.

Also OpenStack files using e.g. $host format for interpolation,
which is not defined in Python configparser documentation.