fix_auth always cleanup all keys
The --v2 option was an optimization to speed up fixing
old encryption keys. It would only fix keys that had not been
upgraded yet.

Unfortunately it is confusing when users are trying to fix
bad v2 keys. It also did not speed up the process that much.

We have updated it across versions, so now the flag has
different meanings on different versions. very confusing.

This removes the flag.