Change caching for OpenstackEventMonitor
OpenstackEventMonitor implementation classes were cached to be sure that we
didn't do expensive connection tests each time events were gathered for an
openstack provider.

However, the cache was a permenant cache and was only cleared when the appliance
was restarted.  The new cache will invalidate every 5 minutes.

This cache invalidation will allow the OpenstackEventMonitor to recover from
communication failures with the AMQP service.

https://bugzilla.redhat.com/show_bug.cgi?id=1247200