Add raw operations to OrchestrationStack and its subclasses
Operations include to create, update, delete, and query status of a stack at its provider.
Upon creating a raw stack, an instance of OrchestrationStack is also inserted to vmdb.

Each provider is required to implement a subclass of OrchestrationStack.

Move create_stack and stack_status methods from CloudManager to OrchestrationStack