Defer loading of both nested children and subclasses
Nested children must be loaded immediately in development, because we
need to use require_dependency in order for reloading to work. But
that's not a concern in production, so there we can defer to
Module#autoload.

As for the STI loader -- instead of loading all the children as soon as
the parent is loaded, we can wait until the moment it matters: when
someone (most likely, ActiveRecord) calls #descendants.

This deferal is particularly valuable, because for most of the relevant
mid-level classes, we actually never do anything that would trigger the
load.