Fix handsoap debug logging in the broker.
Requires https://github.com/ManageIQ/handsoap/pull/1

Prior to the handsoap change, if the client sets the log_header_callback,
but not the log_debug_callback, debug would still be set to true. In the
post-response case, this would cause the XML to be pretty formatted for
no reason at all, which can be incredibly expensive. This new code only
does the message processing if the callback has been set.

To leverage these new savings, we must ensure in the broker that we are
not setting the log_body_callback when not necessary.

Broker RSS memory, post handsoap response, with a large environment on
the initial preload:
Before 564.1MB / After 444.8MB / Savings 119.3MB (~21.1%)