Handle duplicate infra host hostnames
If two hosts have the same hostname they will get assigned
the same database ID and overwrite each other every refresh.
To resolve this in addition to looking up a host by hostname
make sure that what is returned does not have a different
ManagedObjectReference to ensure we aren't overwriting a
different host.

https://bugzilla.redhat.com/show_bug.cgi?id=1266561