Change the criteria when an event is not being raised to automate.
Don't raise the event that is not defined in MiqEventDefinition and the event target is of one of the classes that support policy and alert.
So all events for classes that don't support policy and alert can be raised up to automate without an entry in MiqEventDefinition.
This is the behavior before event switchboard.