For consistent test timings, compile the STI loader before the tests.
Previously, a random test would show up as the slowest test because
it was the first test to hit a missing constant which would cause
autoloading/STI loader to kick in and spend lots of time in this test.

With this change, we should have more constistent test times because
the autoloading magic will be done before any of the tests run.