Fix report result serialization and deserialization
It looks like a bug was introduced in fa9984f845b0 where
`MiqReportResult.report_results=` was only expecting a `MiqReport`
object, whereas `#_async_generate_result` will send it a CSV, plain
text, or PDF. It calls `#to_hash` on the object which results in a `NoMethodError`.

As a quick fix, this commit switches on type on both the setter and
getter so that the object passed in can be serialized and deserialized correctly.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1273275