Remove fallback to https
When ssl is chosen, data should be always send in encrypted form,
fallback to non ssl when ssl is not available breaks it. Also
validation will always return valid, while silenty failing
and switching to non ssl.

Only downside is that it's not backwards compatible. New default
will be SSl without validation, so provider needs to be edited to
what is really supported, otherwise validation and refresh will
fail.

Fixes BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=1286629